♻️ refactor printable qr code view to own component
